// src/Models/api.ts
export interface ProcessHistoryDTO {
    /**
     * 
     * @type {string}
     * @memberof ProcessHistoryDTO
     */
    'action'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessHistoryDTO
     */
    'actionStatus'?: ProcessHistoryDTOActionStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof ProcessHistoryDTO
     */
    'comments'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessHistoryDTO
     */
    'timestamp'?: string;
}

export const ProcessHistoryDTOActionStatusEnum = {
    Started: 'STARTED',
    Pending: 'PENDING',
    Completed: 'COMPLETED',
    Rejected: 'REJECTED',
    Cancelled: 'CANCELLED',
    Approved: 'APPROVED'
} as const;

export type ProcessHistoryDTOActionStatusEnum = typeof ProcessHistoryDTOActionStatusEnum[keyof typeof ProcessHistoryDTOActionStatusEnum];
// ...
export interface ReportDTO {
    /**
     * 
     * @type {string}
     * @memberof ReportDTO
     */
    'username'?: string;
    /**
     * 
     * @type {number}
     * @memberof ReportDTO
     */
    'processInstanceId'?: number;
    /**
     * 
     * @type {string}
     * @memberof ReportDTO
     */
    'processName'?: string;
    /**
     * 
     * @type {string}
     * @memberof ReportDTO
     */
    'startTime'?: string;
    /**
     * 
     * @type {Array<ProcessHistoryDTO>}
     * @memberof ReportDTO
     */
    'processHistoryDTOList'?: Array<ProcessHistoryDTO>;
    /**
     * 
     * @type {string}
     * @memberof ReportDTO
     */
    'currentStatus'?: ReportDTOCurrentStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof ReportDTO
     */
    'processDefinitionName'?: string;
}

export const ReportDTOCurrentStatusEnum = {
    Started: 'STARTED',
    Pending: 'PENDING',
    Completed: 'COMPLETED',
    Rejected: 'REJECTED',
    Cancelled: 'CANCELLED',
    Approved: 'APPROVED'
} as const;

export type ReportDTOCurrentStatusEnum = typeof ReportDTOCurrentStatusEnum[keyof typeof ReportDTOCurrentStatusEnum];